requirements.txt
asyncio
datetime
discord.py~=2.4.0
mysql
mysql-connector-python
siegeapi